requirements.txt
# Core dependencies for the computer-use agent
google-generativeai>=0.8.3
playwright>=1.48.0
asyncio
logging
dataclasses
pathlib
typing-extensions

# Development dependencies (optional)
pytest>=8.0.0
pytest-asyncio>=0.24.0
black>=24.0.0
flake8>=7.0.0
mypy>=1.10.0